more commands lots of changes process or rewriting neural network implementation restructured and cleaned logging framework

+++ src/main/java/gk646/jnet/neuralnetwork/builder/DerivativeActivationFunction.java
@@ -4,16 +4,16 @@ public enum DerivativeActivationFunction {
 
     SIGMOID {
         @Override
-        public float apply(float x) {
-            return (float) (1 / (1 + Math.exp(-x)) * (1 - (1 / (1 + Math.exp(-x)))));
+        public double apply(double x) {
+            return  (1 / (1 + Math.exp(-x)) * (1 - (1 / (1 + Math.exp(-x)))));
         }
     },
     RELU {
         @Override
-        public float apply(float x) {
+        public double apply(double x) {
             return x > 0 ? 1 : 0;
         }
     };
 
-    public abstract float apply(float x);
+    public abstract double apply(double x);
 }
